update Godep commit hashes for docker 1.2.0; removed nsinit from serviced; fix ParseRepositoryTag
